[Java] Remove SnapshotReplication.
